requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
responses==0.10.16
setuptools-git==1.2
six==1.15.0
smls==0.0.15
toml==0.10.1
twilio==6.45.0
urllib3==1.25.10
wrapt==1.12.1
